Rename agent packages to disambiguate runtime from module
internal/agent -> internal/agentengine (the agentic runtime) and the copilotagent module -> agentadapter (its command-module integration). Config keys keep the CopilotAgent* prefix since those map to deployed env vars.
