+++ internal/agentengine/agent.go
@@ -1,4 +1,4 @@
-package agent
+package agentengine
 
 import (
 	"context"
