+++ internal/agentengine/brain.go
@@ -1,4 +1,4 @@
-package agent
+package agentengine
 
 import (
 	_ "embed"
