+++ internal/agentengine/brain_refresh.go
@@ -1,4 +1,4 @@
-package agent
+package agentengine
 
 import (
 	"context"
