+++ internal/agentengine/brain_test.go
@@ -1,4 +1,4 @@
-package agent
+package agentengine
 
 import (
 	"strings"
