+++ internal/agentengine/prompts/brain_guidance.md
