+++ internal/agentengine/role_test.go
@@ -1,4 +1,4 @@
-package agent
+package agentengine
 
 import (
 	"testing"
